requirements.txt
--pre
py-cord==2.0.0rc1
python-dotenv
pydotted
loguru
flask
pandas
watchdog
pyyaml
black
pymongo
dnspython
git+https://github.com/dimitrismistriotis/profanity-check